requirements.txt
fastapi
uvicorn[standard]
openai
twilio
python-dotenv
requests
pandas
jinja2
stripe
itsdangerous
python-multipart
anthropic
google-auth
google-api-python-client
pytz
psycopg2-binary